More Fixing/Setting up IC3
This is a bunch of changes made to models, index, and sv_scripts to clean up code and get everything working. There is an error with adding a new restaurant to the database, primarily with defining a currentId field. Will fix in next update but I'm tired rn and need to go to bed.