filtering and return the data is hard
Couple changes to the data.html. i delete the buttons. dont think we need it. started on the filtering data and tryied to fix some visual css. not finish yet for both. still need some work. Soehow the mainEvent is not firing at all. dont know the reason. need to debug